Add the update profile panel, notification panel, comment section, fix the ui and navigation bugs

+++ client/src/components/CommentItem.js
@@ -2,28 +2,43 @@ import React from "react";
 import classes from "./CommentItem.module.scss";
 import { format } from "timeago.js";
 import Delete from "./svg/Delete";
-import { useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import { deleteComment } from "../store/reducers/data";
 
 const CommentItem = (props) => {
   const { comment } = props;
   const dispatch = useDispatch();
+  const userState = useSelector((state) => state.user);
 
   const deleteCommentHandler = () => {
-    dispatch(
-      deleteComment({
-        commentId: comment._id,
-        screamId: comment.screamId,
-        userId: props.screamUserId,
-        socket: props.socket,
-      })
-    );
+    if (
+      props.comment.userHandle === userState.userId ||
+      userState.userId === props.screamUserId
+    ) {
+      dispatch(
+        deleteComment({
+          commentId: comment._id,
+          screamId: comment.screamId,
+          userId: props.screamUserId,
+          socket: props.socket,
+          userHandle: props.comment.userHandle,
+        })
+      );
+    }
   };
 
   return (
     <div className={`${classes["comment-item"]}`}>
       <div className={`${classes["image-container"]}`}>
-        <img src="/images/no-img.png" alt="profile-icon" />
+        {comment.userImageUrl ? (
+          <img
+            src={comment.userImageUrl}
+            alt="profile-icon"
+            referrerPolicy="no-referrer"
+          />
+        ) : (
+          <img src="/images/no-img.png" alt="profile-icon" />
+        )}
       </div>
 
       <div className={`${classes["lower-container"]}`}>
@@ -33,10 +48,14 @@ const CommentItem = (props) => {
         <div className={classes.timeago}>{format(comment.createdAt)}</div>
         <div className={classes["comment-body"]}>{comment.body}</div>
       </div>
-
-      <div className={classes["comment-actions"]}>
-        <Delete onClick={deleteCommentHandler} />
-      </div>
+      {props.comment.userHandle === userState.userId ||
+      userState.userId === props.screamUserId ? (
+        <div className={classes["comment-actions"]}>
+          <Delete onClick={deleteCommentHandler} />
+        </div>
+      ) : (
+        <></>
+      )}
     </div>
   );
 };
